Rename: Move response dto
