+++ server/src/main/java/project/manager/server/controller/RegionController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import project.manager.server.dto.ResponseDto;
+import project.manager.server.dto.reponse.ResponseDto;
 import project.manager.server.dto.reponse.region.GuDto;
 import project.manager.server.dto.reponse.region.SiDto;
 import project.manager.server.service.RegionService;
@@ -25,13 +25,13 @@ public class RegionController {
     @GetMapping("/si")
     public ResponseDto<List<SiDto>> showSiList() {
 
-        return new ResponseDto<List<SiDto>>(regionService.readSiList());
+        return new ResponseDto<>(regionService.readSiList());
     }
 
     //구 목록 읽어오기
     @GetMapping("/gu/{siId}")
     public ResponseDto<List<GuDto>> showGuList(@PathVariable Long siId) {
 
-        return new ResponseDto<List<GuDto>>(regionService.readGuList(siId));
+        return new ResponseDto<>(regionService.readGuList(siId));
     }
 }
