+++ server/src/main/java/project/manager/server/controller/UserController.java
@@ -7,7 +7,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import project.manager.server.dto.ResponseDto;
+import project.manager.server.dto.reponse.ResponseDto;
 import project.manager.server.dto.reponse.UserDto;
 import project.manager.server.dto.request.UserRequestDto;
 import project.manager.server.dto.request.report.UserReportRequestDto;
