+++ server/src/main/java/project/manager/server/controller/manage/BuildingReportController.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 
 import org.springframework.web.bind.annotation.*;
 
-import project.manager.server.dto.ResponseDto;
+import project.manager.server.dto.reponse.ResponseDto;
 import project.manager.server.service.post.building.BuildingPostService;
 import project.manager.server.service.report.BuildingReportService;
