+++ server/src/main/java/project/manager/server/controller/post/ApplyController.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 
 import org.springframework.web.bind.annotation.*;
 
-import project.manager.server.dto.ResponseDto;
+import project.manager.server.dto.reponse.ResponseDto;
 import project.manager.server.dto.request.post.building.ApplyRequestDto;
 import project.manager.server.service.post.building.ApplyService;
