+++ server/src/main/java/project/manager/server/controller/post/PostTypeController.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 
 import org.springframework.web.bind.annotation.*;
 
-import project.manager.server.dto.ResponseDto;
+import project.manager.server.dto.reponse.ResponseDto;
 import project.manager.server.dto.request.post.contest.PostTypeRequestDto;
 import project.manager.server.service.post.contest.*;
